+++ src/components/orders-feed/feed-orders-detalis/feed-order-details.tsx
@@ -1,0 +1,5 @@
+export const FeedOrderDetails = () =>{
+    return(
+        <div>кря-кря</div>
+    )
+}
