+++ src/components/orders-feed/orders-feed.module.css
@@ -20,3 +20,6 @@
     align-items: center;
 }
 
+a.link {
+  color: white;
+}
