+++ src/pages/feed/feed.module.css
@@ -32,6 +32,3 @@
     border-radius: 6px;
   }
 
-  .stats{
-   
-  }
